parsers/requirements.txt
fastapi
lxml
python-dotenv
Requests
uvicorn
xmltodict
